+++ tests/tests/index.js
@@ -189,6 +189,7 @@ import './es.weak-set';
 
 QUnit.module('ESNext');
 import './esnext.aggregate-error';
+import './esnext.array.is-template-object';
 import './esnext.array.last-item';
 import './esnext.array.last-index';
 import './esnext.composite-key';